+++ massiv/tests/Data/Massiv/Array/MutableSpec.hs
@@ -7,6 +7,7 @@
 module Data.Massiv.Array.MutableSpec (spec) where
 
 import Control.Concurrent.Async
+import Control.Monad (when)
 import Control.Monad.ST
 import Data.Functor.Identity
 import Data.List as L
@@ -43,6 +44,50 @@ prop_generateMakeIO _ _ (Arr arr) = monadicIO $ do
   arr' <- run $ generateArray (getComp arr) (size arr) (evaluateM arr)
   return (arr === arr')
 
+prop_shrinkIO ::
+     ( Mutable r ix Int
+     , Resize r ix
+     , Source r Ix1 Int
+     )
+  => r
+  -> Proxy ix
+  -> ArrIx r ix Int
+  -> Property
+prop_shrinkIO _ _ (ArrIx arr ix) =
+  monadicIO $
+  run $ do
+    marr <- thaw arr
+    sarr <- unsafeFreeze (getComp arr) =<< unsafeLinearShrink marr (Sz ix)
+    pure (A.foldlS (.&&.) (property True) $ A.zipWith (===) (flatten arr) (flatten sarr))
+
+prop_growShrinkIO ::
+     ( Show (Array r ix Int)
+     , Eq (Array r ix Int)
+     , Mutable r ix Int
+     , Extract r ix Int
+     , Num ix
+     , Load (EltRepr r ix) ix Int
+     )
+  => r
+  -> Proxy ix
+  -> Arr r ix Int
+  -> NonNegative Int
+  -> Property
+prop_growShrinkIO _ _ (Arr arr) (NonNegative delta) =
+  monadicIO $
+  run $ do
+    marr <- thaw arr
+    let sz = size arr
+    k <- getDimM (unSz sz) (dimensions sz)
+    -- increase the outer most dimension, just so the structure doesn't change
+    newSz <- Sz <$> setDimM (unSz sz) (dimensions sz) (k + delta)
+    gMarr <- unsafeLinearGrow marr newSz
+    -- Make sure we can write into the newly allocated area
+    when (delta > 0) $ write' gMarr (unSz newSz - 1) delta
+    garr <- compute . extract' 0 sz <$> unsafeFreeze (getComp arr) gMarr
+    sarr <- freeze (getComp arr) =<< unsafeLinearShrink gMarr sz
+    pure (garr === arr .&&. sarr === arr)
+
 prop_atomicModifyIntArrayMany :: ArrIx P Ix2 Int -> Array B Ix1 Int -> Property
 prop_atomicModifyIntArrayMany (ArrIx arr ix) barr =
   monadicIO $ do
@@ -58,15 +103,16 @@ prop_atomicModifyIntArrayMany (ArrIx arr ix) barr =
 
 
 prop_atomicReadIntArrayMany :: Array P Ix2 Int -> Array B Ix1 Ix2 -> Property
-prop_atomicReadIntArrayMany arr bix = monadicIO $ do
+prop_atomicReadIntArrayMany arr bix =
+  monadicIO $
   run $ do
-      marr <- thaw arr
-      as :: Array N Ix1 (Maybe Int) <- forM bix (A.read marr)
-      as' <- forM bix (atomicReadIntArray marr)
-      pure (as === as')
+    marr <- thaw arr
+    as :: Array N Ix1 (Maybe Int) <- forM bix (A.read marr)
+    as' <- forM bix (atomicReadIntArray marr)
+    pure (as === as')
 
 
-prop_atomicWriteIntArrayMany :: Array P Ix2 Int -> Array B Ix1 Ix2 -> (Fun Ix2 Int) -> Property
+prop_atomicWriteIntArrayMany :: Array P Ix2 Int -> Array B Ix1 Ix2 -> Fun Ix2 Int -> Property
 prop_atomicWriteIntArrayMany arr bix f =
   monadicIO $
   run $ do
@@ -96,20 +142,30 @@ prop_unfoldrReverseUnfoldl sz1 f i =
           rev a = computeAs P $ backpermute' sz1 (\ix1 -> unSz sz1 - ix1 - 1) a
 
 
+
 mutableSpec ::
      ( Show r
-     , Show (Array r Ix3 Int)
      , Show (Array r Ix1 Int)
      , Show (Array r Ix2 Int)
-     , Eq (Array r Ix3 Int)
+     , Show (Array r Ix3 Int)
      , Eq (Array r Ix1 Int)
      , Eq (Array r Ix2 Int)
-     , Mutable r Ix3 Int
+     , Eq (Array r Ix3 Int)
      , Mutable r Ix1 Int
      , Mutable r Ix2 Int
-     , Construct r Ix3 Int
+     , Mutable r Ix3 Int
      , Construct r Ix1 Int
      , Construct r Ix2 Int
+     , Construct r Ix3 Int
+     , Extract r Ix1 Int
+     , Extract r Ix2 Int
+     , Extract r Ix3 Int
+     , Resize r Ix1
+     , Resize r Ix2
+     , Resize r Ix3
+     , Load (EltRepr r Ix1) Ix1 Int
+     , Load (EltRepr r Ix2) Ix2 Int
+     , Load (EltRepr r Ix3) Ix3 Int
      )
   => r
   -> SpecWith ()
@@ -131,7 +187,14 @@ mutableSpec r =
       it "Ix1" $ property $ prop_generateMakeIO r (Proxy :: Proxy Ix1)
       it "Ix2" $ property $ prop_generateMakeIO r (Proxy :: Proxy Ix2)
       it "Ix3" $ property $ prop_generateMakeIO r (Proxy :: Proxy Ix3)
-
+    describe "shrink" $ do
+      it "Ix1" $ property $ prop_shrinkIO r (Proxy :: Proxy Ix1)
+      it "Ix2" $ property $ prop_shrinkIO r (Proxy :: Proxy Ix2)
+      it "Ix3" $ property $ prop_shrinkIO r (Proxy :: Proxy Ix3)
+    describe "grow+shrink" $ do
+      it "Ix1" $ property $ prop_growShrinkIO r (Proxy :: Proxy Ix1)
+      it "Ix2" $ property $ prop_growShrinkIO r (Proxy :: Proxy Ix2)
+      it "Ix3" $ property $ prop_growShrinkIO r (Proxy :: Proxy Ix3)
 
 generateSpec :: Spec
 generateSpec = do